requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.22.0
pillow==10.3.0
requests==2.32.3
shortuuid==1.0.13
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
urllib3==2.2.2
